[cfe] Remove language version from unreleased experiment errors.
Unreleased experiments have no language version even though we pretend
they have the current language version. However, it's problematic to
mention the current language version in the error messages as the errors
go into the expectation files. It's a problem that we need to update an
unknowable amount of files whenever we bump the version, so it's much
better to omit the current language version from these error messages.

This change is part of the effort to make it possible to easily bump
tools/VERSION without unexpected issues whenever tests have to have
their pinned language versions updated as well.

Bug: b/30474526
Change-Id: Ia58b6ae002000511cea98ed83d256505deb986dd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/330108
Commit-Queue: Jonas Termansen <sortie@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>