[vm] Remove dead deferred loading handling for object pools.
When each Code had its own ObjectPool, and we were loading a deferred unit, the ObjectPool would be initialized before PostLoad for some Instance potentially canonicalized the entry to another object loaded in some other unit that was not a dominator of the current loading unit. We now are always only in bare-metal mode with a global object pools. The equivalent reloading of pool entries now happens in UnitDeserializationRoots::ReadRoots.

TEST=ci
Change-Id: I021cd759abf25dfb90494a1d6334e5d4a1173ac4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/331120
Reviewed-by: Daco Harkes <dacoharkes@google.com>
Commit-Queue: Ryan Macnak <rmacnak@google.com>